+++ Core/Src/motor.h
@@ -19,7 +19,7 @@
 // 电机控制函数声明
 void Car_test(uint8_t speed);
 void Motor_Init(void);
-void Motor_SetSpeed(uint8_t motor, uint8_t speed);
+void Motor_SetSpeed(uint8_t motor, uint16_t speed);
 void Motor_SetDirection(uint8_t motor, uint8_t direction);
 void Motor_Control(uint8_t motor, uint8_t direction, int speed);
 void Motor_Stop(uint8_t motor);
@@ -27,9 +27,9 @@ void Motor_StopAll(void);
 
 // 小车运动控制函数
 void Car_Forward(uint16_t speed);
-void Car_Backward(uint8_t speed);
-void Car_TurnLeft(uint8_t speed);
-void Car_TurnRight(uint8_t speed);
+void Car_Backward(uint16_t speed);
+void Car_TurnLeft(uint16_t speed);
+void Car_TurnRight(uint16_t speed);
 void Car_Stop(void);
 
 #endif //FINAL_FINA_MOTOR_H
